Update User handle

+++ app/delete_comment.py
@@ -11,8 +11,11 @@ class DeleteComment(BlogHandler):
             ## Getting comment key using comment_id
             key = db.Key.from_path('Comment', int(comment_id),
                                    parent=blog_key())
-            c = db.get(key)            
-            c.delete()
-            self.redirect("/"+post_id)
+            if c.author == int(self.read_secure_cookie('user_id')):
+                c = db.get(key)            
+                c.delete()
+                self.redirect("/"+post_id)
+            else:
+                self.redirect("/")    
         else:
             self.redirect("/")
